starRate에 short를 int로 바꿈

+++ hotel/src/main/java/com/mjc/hotel/hotels/HotelDto.java
@@ -24,7 +24,7 @@ public class HotelDto extends BaseDto implements IHotel{
     private String email;
     private LocalTime checkIn;
     private LocalTime checkOut;
-    private Short starRate;
+    private int starRate;
     private Boolean isActive;
     private String latitude;
     private String longitude;
